requirements.txt
tensorflow
Cython>=0.19.2
numpy>=1.7.1
matplotlib>=1.3.1
Pillow>=2.3.0
six>=1.1.0
tqdm